Move YouTube event emitters to their own module
